+++ ui/src/components/os/index.ts
@@ -4,6 +4,7 @@ export { default as StatusIndicator } from './StatusIndicator'
 export { default as ApplicationPanel } from './ApplicationPanel'
 export { default as KernelStatusPanel } from './KernelStatusPanel'
 export { default as FilesystemPanel } from './FilesystemPanel'
+export { default as LogsPanel } from './LogsPanel'
 export { panels } from './panels'
 export { formatUptime, formatBytes } from '../../utils/osUtils'
 export type {
@@ -12,4 +13,5 @@ export type {
   DirEntry,
   PanelType,
   StatusType,
+  LogEntry,
 } from '../../types/osTypes'